+++ scripts/resolve-did.js
@@ -1,15 +1,13 @@
 import dotenv from 'dotenv';
-import { packageStore } from '../src/packageAgent.ts';
-
+import { getResolver } from '@verida/vda-did-resolver';
 dotenv.config();
 
 const testnetDID = process.env.TESTNET_DID;
 
 (async () => {
-    const { agent } = await packageStore.initialize();
-
+    const resolver = getResolver();
     try {
-        const did = await agent.resolveDid({ didUrl: testnetDID});
+        const did = await resolver.vda(testnetDID);
         console.log(did);
     } catch (error) {
         console.error("❌ Error resolving DID:", error);