+++ src/packageAgent.ts
@@ -14,18 +14,20 @@ import path from 'path';
 import { fileURLToPath } from 'url';
 import { co2 } from "@tgwf/co2";
 import { getResolver } from "@verida/vda-did-resolver";
+import { DIDClient } from '@verida/did-client';
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
 
 dotenv.config();
 
-const universalResolver = getUniversalResolverFor(['cheqd', 'key', 'vda']);
+const universalResolver = getUniversalResolverFor(['cheqd', 'key']);
 const vdaResolver = getResolver();
 
 const packageJsonPath = path.join(__dirname, '../package.json');
 
 let cheqdMainnetProvider: CheqdDIDProvider | null = null;
+let cheqdTestnetProvider: CheqdDIDProvider | null = null;
 let packageAgent: IOVAgent | null = null;
 let currentDIDKey: string | null = null;
 let signedVCs: VerifiableCredential[] = [];
@@ -38,9 +40,9 @@ const initializePackageAgent = async ({ payerSeed, didRecoveryPhrase }: { payerS
     let cosmosPayerSeed = payerSeed || process.env.COSMOS_PAYER_SEED || '';
     let didMnemonic = didRecoveryPhrase || process.env.PACKAGE_DID_RECOVERY_PHRASE || '';
 
-    cheqdMainnetProvider = createCheqdProvider(CheqdNetwork.Mainnet, cosmosPayerSeed, process.env.CHEQD_RPC_URL || 'https://cheqd.originvault.box:443');
-
-    packageAgent = createOVAgent(cheqdMainnetProvider, universalResolver, vdaResolver);
+    cheqdMainnetProvider = createCheqdProvider(CheqdNetwork.Mainnet, cosmosPayerSeed, process.env.CHEQD_MAINNET_RPC_URL || 'https://cheqd.originvault.box:443');
+    cheqdTestnetProvider = createCheqdProvider(CheqdNetwork.Testnet, cosmosPayerSeed, process.env.CHEQD_TESTNET_RPC_URL || 'https://rpc.cheqd.network');
+    packageAgent = createOVAgent(cheqdMainnetProvider, universalResolver, vdaResolver, cheqdTestnetProvider);
 
     if(!packageAgent) {
         throw new Error("Package agent could not be initialized");
@@ -211,7 +213,7 @@ const initializePackageAgent = async ({ payerSeed, didRecoveryPhrase }: { payerS
         return result;
     }
 
-    return { agent: packageAgent, did: packageJsonDIDString, key: currentDIDKey, credentials: signedVCs, publishWorkingKey, publishRelease, privateKeyStore };
+    return { agent: packageAgent, did: packageJsonDIDString, key: currentDIDKey, credentials: signedVCs, publishWorkingKey, publishRelease, privateKeyStore, cheqdTestnetProvider, cheqdMainnetProvider };
 }
 
 const packageStore: AgentStore = {